requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
fastapi==0.115.3
fb-idb==1.1.7
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
mysqlclient==2.2.5
protobuf==5.28.2
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.16
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
treelib==1.7.0
typing_extensions==4.12.2
uvicorn==0.32.0
